Moved SN client and servers into their own files
