+++ cli/SingleNodeClient.hpp
@@ -17,6 +17,9 @@
  * under the License.
  **/
 
+#ifndef QUICKSTEP_CLI_SINGLE_NODE_CLIENT_HPP_
+#define QUICKSTEP_CLI_SINGLE_NODE_CLIENT_HPP_
+
 #include <grpc++/grpc++.h>
 
 #include <iostream>
@@ -34,23 +37,15 @@ using grpc::Channel;
 using grpc::ClientContext;
 using grpc::Status;
 
-static bool ValidatePort(const char *flagname, std::int32_t value) {
-  if (value > 0 && value < 65536) {
-    return true;
-  }
-  std::printf("Invalid value for --%s: %d\n", flagname, static_cast<int>(value));
-  return false;
-}
-
-DEFINE_int32(port, 3000, "Listens for TCP connection on this port.");
-DEFINE_validator(port, &ValidatePort);
-
 namespace quickstep {
 
-class QuickstepClient {
+/**
+ * A simple wrapper class used to do CLI interactions with QuickstepCLI via the gRPC interface.
+ */
+class SingleNodeClient {
  public:
-  explicit QuickstepClient(std::shared_ptr<Channel> const &channel)
-      : stub_(SingleNodeServerRequest::NewStub(channel)) {}
+  explicit SingleNodeClient(std::shared_ptr<Channel> const &channel)
+    : stub_(SingleNodeServer::NewStub(channel)) {}
 
   /**
    * Assembles the client's payload, sends it and presents the response back from the server.
@@ -87,22 +82,9 @@ class QuickstepClient {
     return response.query_result();
   }
 
-  std::unique_ptr<SingleNodeServerRequest::Stub> stub_;
+  std::unique_ptr<SingleNodeServer::Stub> stub_;
 };
 
 }  // namespace quickstep
 
-int main(int argc, char **argv) {
-  google::InitGoogleLogging(argv[0]);
-  gflags::ParseCommandLineFlags(&argc, &argv, true);
-
-  // Attempts to send a single query retrieved from stdin to the Quickstep Server.
-  quickstep::QuickstepClient qs_client(
-    grpc::CreateChannel("localhost:" + std::to_string(FLAGS_port),
-                        grpc::InsecureChannelCredentials()));
-  std::string user_query;
-  std::cin >> user_query;
-  std::cout << "query: " << user_query << std::endl;
-  std::cout << qs_client.SendQuery(user_query) << std::endl;
-  return 0;
-}
+#endif //QUICKSTEP_CLI_SINGLE_NODE_CLIENT_HPP_
